// src/modules/cognitive-problem/domain/enums/cognitive-problem-category.enum.ts
// domain/enums/cognitive-problem-category.enum.ts
export enum CognitiveProblemCategory {
  DESIGN = 'design',
  FUNCTIONALITY = 'functionality',
  NAVIGATION = 'navigation',
  CONTENT = 'content',
  PERFORMANCE = 'performance',
  ACCESSIBILITY = 'accessibility',
  USABILITY = 'usability',
  OTHER = 'other',
}